requirements.txt
fastapi[standard]
pydantic
PyJWT
